move buttons to own namespace
