+++ src/main/java/com/lothrazar/cyclicmagic/gui/GuiWandInventory.java
@@ -2,6 +2,10 @@ package com.lothrazar.cyclicmagic.gui;
 
 import org.lwjgl.opengl.GL11;
 import com.lothrazar.cyclicmagic.Const;
+import com.lothrazar.cyclicmagic.gui.button.ButtonBuildToggle;
+import com.lothrazar.cyclicmagic.gui.button.ButtonPassiveToggle;
+import com.lothrazar.cyclicmagic.gui.button.ButtonRecharge;
+import com.lothrazar.cyclicmagic.gui.button.ButtonSpellCircle;
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.util.ResourceLocation;
