+++ src/main/java/com/lothrazar/cyclicmagic/gui/button/ButtonSpell.java
@@ -1,4 +1,4 @@
-package com.lothrazar.cyclicmagic.gui;
+package com.lothrazar.cyclicmagic.gui.button;
 
 import java.util.ArrayList;
 import java.util.List;
